Controller, Service, Impl created

+++ src/main/java/ru/skypro/testcalculator/service/impl/CalculatorServiceImpl.java
@@ -22,6 +22,6 @@ public class CalculatorServiceImpl implements CalculatorService {
 
     @Override
     public double divide(int num1, int num2) {
-        return num1 / num2;
+        return (double) num1 / num2;
     }
 }
